requirements.txt
python_dateutil >= 2.5.3
setuptools >= 21.0.0
urllib3==2.2.2
pydantic >= 2
typing-extensions >= 4.7.1
aiohttp >= 3.0.0
aiohttp-retry >= 2.8.3
python-dotenv >= 1.0.1